logo-designer/requirements.txt
requests>=2.31.0
reportlab>=4.0.0
cairosvg>=2.7.0
Pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
